Refactor MonitoringService to use model_dump for saving configuration and add comprehensive tests for monitoring functionality
